# sstable: reduce writer allocations during index flushing